+++ app/services/stub_init_service.py
@@ -1,7 +1,7 @@
 import datetime as dt
 
 from models import ClientModel, EmployeeModel
-from models.user_model import EmployeeInSchema
+from schemas.user_schemas import EmployeeInSchema
 from models.choices import Gender
 from services.base_service import BaseService
 from services.client_service import ClientService
