+++ app/tests/test_user.py
@@ -6,7 +6,7 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 from sqlalchemy import select
 from fastapi import status
 from models import EmployeeModel
-from models.user_model import EmployeeInSchema, EmployeeInOptionalSchema
+from schemas.user_schemas import EmployeeInSchema, EmployeeInOptionalSchema
 from models.choices import Gender
 from services.stub_init_service import LAST_NAMES
 from tests.conftest import engine
